Conflicts:
	palladian-core/src/main/java/ws/palladian/classification/utils/ClassificationUtils.java
	palladian-core/src/main/java/ws/palladian/extraction/entity/dataset/WikipediaDatasetCreator.java
	palladian-core/src/main/java/ws/palladian/extraction/entity/tagger/AlchemyNer.java
	palladian-core/src/main/java/ws/palladian/extraction/location/LocationExtractorUtils.java
	palladian-core/src/main/java/ws/palladian/extraction/location/YahooLocationExtractor.java
	palladian-core/src/main/java/ws/palladian/extraction/location/disambiguation/FeatureExtractorSetting.java
	palladian-core/src/main/java/ws/palladian/extraction/location/evaluation/TudLoc2013DatasetIterable.java
	palladian-core/src/main/java/ws/palladian/extraction/location/sources/CollectionLocationStore.java
	palladian-core/src/main/java/ws/palladian/extraction/location/sources/importers/GeonamesImporter.java
	palladian-experimental/src/main/java/ws/palladian/extraction/location/persistence/LuceneLocationSource.java
	palladian-retrieval/src/main/java/ws/palladian/retrieval/HttpRequest.java
	palladian-retrieval/src/main/java/ws/palladian/retrieval/ranking/services/BitlyClicks.java
	palladian-retrieval/src/main/java/ws/palladian/retrieval/ranking/services/FacebookLinkStats.java
	palladian-retrieval/src/main/java/ws/palladian/retrieval/search/SearcherFactory.java
	palladian-retrieval/src/main/java/ws/palladian/retrieval/search/events/EventbriteSearcher.java
	palladian-retrieval/src/main/java/ws/palladian/retrieval/search/events/EventfulSearcher.java
	palladian-retrieval/src/main/java/ws/palladian/retrieval/search/web/BlekkoSearcher.java

Former-commit-id: c00e11be5b7bd08becc67c74f6c7d62473db5fa0

+++ palladian-commons/src/main/java/ws/palladian/helper/math/FatStats.java
@@ -1,6 +1,10 @@
 package ws.palladian.helper.math;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.List;
 
 import org.apache.commons.lang.Validate;
 
@@ -191,7 +195,7 @@ public class FatStats implements Stats {
         }
         List<Double> temp = getDoubleValues();
         Collections.sort(temp);
-        double n = (double)(p / 100.) * temp.size();
+        double n = p / 100. * temp.size();
         if (n == (int)n) {
             return 0.5 * temp.get((int)n - 1) + 0.5 * temp.get((int)n);
         } else {
